chore: refactor app.js stack screen code
